+++ utilities.h
@@ -3,6 +3,6 @@
 
 char *itoa(unsigned int);
 
-void clear_buffer(char*, unsigned int);
+void clear_buffer(char*, size_t);
 
 #endif
